add: add more modularity in the code
