+++ newscript.ts
@@ -3,15 +3,19 @@ import { htmlElement } from "./modularMethods/CommonElements.js";
 import { opInclude } from "./modularMethods/booleanOperation.js";
 import {
   handleMC,
-  handleClosingHistorySection,
+  handleMR,
+  handleMS,
+  handleMplusAndMinus,
 } from "./modularMethods/OperationsRelatedToMemory.js";
+
+import { handleHistory } from "./modularMethods/OperationsRelatedToHistory.js";
 
 import {
   replaceAll,
   degreeRadianChange,
 } from "./modularMethods/ReplacingStringOperation.js";
 
-class Calculator {
+export class Calculator {
   validKeyboardCharacters = [
     "0",
     "1",
@@ -46,6 +50,10 @@ class Calculator {
     this.calculatorInput.value = ``;
     this.regex = /d/;
     this.darkLightFlag = 0;
+    this.initiateEventListener();
+  }
+
+  initiateEventListener() {
     htmlElement.calculatorButtons.addEventListener("click", (event) => {
       this.handleClickOnCalculator(event);
     });
@@ -87,20 +95,20 @@ class Calculator {
     });
 
     document.getElementById("mr-button")!.addEventListener("click", () => {
-      this.handleMR();
+      handleMR.call(this);
     });
 
     document.querySelectorAll("#M-plus-minus-button").forEach((element) => {
       element.addEventListener("click", (event) => {
-        this.handleMplusAndMinus(event.target as EventTarget);
+        handleMplusAndMinus.call(this, event.target as EventTarget);
       });
     });
     document.getElementById("ms-button")!.addEventListener("click", () => {
-      this.handleMS();
+      handleMS.call(this);
     });
 
     htmlElement.viewHistoryButton.addEventListener("click", () => {
-      this.handleHistory();
+      handleHistory.call(this);
     });
 
     htmlElement.darkLightButton.addEventListener("click", () => {
@@ -391,71 +399,6 @@ class Calculator {
     }
   }
 
-  //added to handle single memory operation
-  handleMR() {
-    let val = localStorage.getItem("calculationOutput");
-    if (val) this.calculatorInput.value += val;
-  }
-
-  handleMplusAndMinus(ref: EventTarget) {
-    console.log("function called");
-
-    let previousOutput = localStorage.getItem("calculationOutput");
-    let num = eval(
-      this.resultFuncInitialEvaluation(this.calculatorInput.value)
-    );
-    let classname: string = "";
-    if ("className" in ref) {
-      classname = ref.className as string;
-    }
-
-    let val =
-      classname === "plus"
-        ? `${+num + +previousOutput!}`
-        : `${+num - +previousOutput!}`;
-
-    localStorage.setItem("calculationOutput", val);
-    this.calculatorInput.value = val;
-  }
-
-  handleMS() {
-    let num = eval(
-      this.resultFuncInitialEvaluation(this.calculatorInput.value)
-    );
-    localStorage.setItem("calculationOutput", num);
-  }
-
-  //added to handle the scenario where user clicks on see history button
-  handleHistory() {
-    (htmlElement.textBoxInput as HTMLInputElement).style.display = "none";
-    (htmlElement.advanceOperationsSection as HTMLDivElement).style.display =
-      "none";
-    htmlElement.calculatorButtons.style.display = "none";
-
-    let closingHistoryButton = document.createElement("img");
-    closingHistoryButton.setAttribute(
-      "class",
-      "imageHW hamburger-menu close-history"
-    );
-    if (this.darkLightFlag === 1) {
-      closingHistoryButton.setAttribute("src", "../images/cross.png");
-    } else {
-      closingHistoryButton.setAttribute("src", "../images/x.png");
-    }
-
-    closingHistoryButton.setAttribute("alt", "closing history section button");
-    closingHistoryButton.addEventListener("click", () => {
-      handleClosingHistorySection();
-    });
-    htmlElement.enclosingCalculatorDiv.appendChild(closingHistoryButton);
-
-    let newElement = document.createElement("div");
-    newElement.setAttribute("class", "history-section");
-    htmlElement.enclosingCalculatorDiv.appendChild(newElement);
-
-    this.showHistoryContent();
-  }
-
   //added to handle the scenario when user clicks on dark to light mode
   handleDarkLightMode() {
     htmlElement.darkLightButton.classList.toggle("selected");
@@ -522,59 +465,7 @@ class Calculator {
         });
     }
   }
-  //added to show the basic history content in the history page
-  showHistoryContent() {
-    document.getElementsByClassName(
-      "history-section"
-    )[0].innerHTML = `<div class = "history-title-clear-button">
-          <div class = "history-title">History</div>
-          <div class = "clear-history-button">
-          
-          <button class = "clear-history-button">Clear</button>
-          </div>
-          </div>`;
-    let arr = JSON.parse(localStorage.getItem("history-array")!);
-    let historyHtmlCode = "";
-    for (let i = arr.length - 1; i >= 0; i--) {
-      historyHtmlCode += `<div class = "history-data" id = "${arr[i][1]}">
-              <div>${arr[i][0]} &nbsp;= &nbsp;</div>
-              <div>${arr[i][1]} </div>
-          </div>`;
-    }
-    document.getElementsByClassName("history-section")[0].innerHTML +=
-      historyHtmlCode;
 
-    document.querySelectorAll(".history-data").forEach((element) => {
-      element.addEventListener("click", (event: Event) => {
-        if ("id" in event.target!)
-          this.handleClickOnHistoryData(event.target.id as string);
-      });
-    });
-
-    document
-      .getElementsByClassName("clear-history-button")[0]
-      .addEventListener("click", () => {
-        this.handleClearHistory();
-      });
-  }
-
-  //handle the scenario when user clicks on clear button in the see history page
-  handleClearHistory() {
-    localStorage.setItem("history-array", JSON.stringify([]));
-    this.showHistoryContent();
-  }
-
-  //added to implement feature -  user clicks on a particular history the data is automatically added to calculator input
-  handleClickOnHistoryData(value: string) {
-    handleClosingHistorySection();
-    this.regex = /(\d+)$/;
-    if (this.regex.test(this.calculatorInput.value)) {
-      this.calculatorInput.value += `X`;
-      this.calculatorInput.value += `${value}`;
-    } else {
-      this.calculatorInput.value += `${value}`;
-    }
-  }
   //added to induce keyboard input feature
   handleKeyboardInput(event: Event) {
     let keyValue: string = "";
